add job and profile nested routes
