+++ src/app/jobs/[id]/page.tsx
@@ -1,0 +1,20 @@
+import { getJob } from '@/lib/utils'
+
+export const revalidate = 3600 // revalidate the data at most every hour
+
+export default async function Page({
+    params: { id },
+}: {
+    params: { id: string }
+}) {
+    const job = await getJob(parseInt(id))
+
+    if (!job) return <h1>Whoopsie, we cant find that job rn.</h1>
+
+    return (
+        <main>
+            <h1>{job.postedBy.username}</h1>
+
+        </main>
+    )
+}
