+++ src/lib/utils.ts
@@ -1,21 +1,21 @@
 import { type ClassValue, clsx } from "clsx"
 import { twMerge } from "tailwind-merge"
-import {cache} from "react" 
+import { cache } from "react"
 import { PrismaClient } from '@prisma/client';
 
 const prisma = new PrismaClient();
 
- 
+
 export function cn(...inputs: ClassValue[]) {
   return twMerge(clsx(inputs))
 }
 
-export const getJobs = cache(async() => {
+export const getJobs = cache(async () => {
   const jobs = await prisma.job.findMany()
-  return jobs  
+  return jobs
 })
 
-export const getProfile = cache(async(id: number) => {
+export const getProfile = cache(async (id: number) => {
   const profile = await prisma.account.findUnique({
     where: {
       id
@@ -24,10 +24,13 @@ export const getProfile = cache(async(id: number) => {
   return profile
 })
 
-export const getJob = cache(async(id: number) => {
+export const getJob = cache(async (id: number) => {
   const job = await prisma.job.findUnique({
     where: {
       id
+    },
+    include: {
+      postedBy: true
     }
   })
   return job
